Refactor graph update orchestration into runtime module
Move update/caching logic from main.py into node_editor/graph_runtime.py and introduce GraphRuntime state ownership for async/sync loops. Keep update_node_info API available from the new module and add focused unit coverage for runtime state persistence.
